[S-2] 홈화면 모임목록 렌더링 에러를 해결하라 close #27 모임목록 서버데이터 as-is한 작업을 반영한 뒤에 생기는 에러를 해결합니다.
우선 화면을 새로고침해도 렌더링되기 위해 useSetMeetingList훅 대신 컴포넌트UI에서 서버데이터를 요청한 뒤, useEffect훅으로 스토어 객체에 저장합니다.

그 밖에 서버데이터, 스토어 객체 둘 중 빠른걸로 렌더링하기 위해 조건식을 추가했습니다.

그리고 serializable에러를 해결하기 위해 configureStore객체의 두 번째 인자로 middleware를 전달했습니다.
