+++ src/store.ts
@@ -1,4 +1,4 @@
-import { configureStore } from '@reduxjs/toolkit';
+import { configureStore, getDefaultMiddleware } from '@reduxjs/toolkit';
 
 import authReducer from './modules/authSlice';
 import homeReducer from './modules/homeSlice';
@@ -8,6 +8,9 @@ const store = configureStore({
     auth: authReducer,
     home: homeReducer,
   },
+  middleware: getDefaultMiddleware({
+    serializableCheck: false,
+  }),
 });
 
 export default store;
